Разбил маппер на отдельные классы-мапперы.
