+++ app/src/main/java/ru/androidschool/intensiv/data/mappers/TvShowMapper.kt
@@ -1,0 +1,13 @@
+package ru.androidschool.intensiv.data.mappers
+
+import ru.androidschool.intensiv.data.dbo.TvShowDBO
+import ru.androidschool.intensiv.data.dto.TvShowDetails
+
+object TvShowMapper {
+    fun toTvShowDBO(tvShowDetails: TvShowDetails): TvShowDBO {
+        return TvShowDBO(
+            id = tvShowDetails.id,
+            name = tvShowDetails.name
+        )
+    }
+}
