+++ app/src/main/java/ru/androidschool/intensiv/presentation/search/SearchFragment.kt
@@ -16,14 +16,14 @@ import kotlinx.android.synthetic.main.fragment_search.*
 import kotlinx.android.synthetic.main.progress_bar.*
 import kotlinx.android.synthetic.main.search_toolbar.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.vo.MovieVO
+import ru.androidschool.intensiv.data.mappers.MovieMapper
 import ru.androidschool.intensiv.data.network.MovieApiClient
+import ru.androidschool.intensiv.data.vo.MovieVO
+import ru.androidschool.intensiv.presentation.feed.MovieItem
 import ru.androidschool.intensiv.ui.applyProgressBar
 import ru.androidschool.intensiv.ui.applySchedulers
-import ru.androidschool.intensiv.presentation.feed.MovieItem
 import ru.androidschool.intensiv.ui.onTextChangedPublishSubject
 import ru.androidschool.intensiv.utils.Const
-import ru.androidschool.intensiv.data.mappers.MovieFinderAppConverter
 import ru.androidschool.intensiv.utils.ViewFeature
 import timber.log.Timber
 import java.util.concurrent.TimeUnit
@@ -100,7 +100,7 @@ class SearchFragment : Fragment(R.layout.fragment_search) {
                 { // в случае успешного получения данных:
                     response ->
                     val movieResultList = response.results
-                    val movieList = movieResultList.map { movie -> MovieFinderAppConverter.toMovieVO(movie, ViewFeature.SEARCHED) }
+                    val movieList = movieResultList.map { movie -> MovieMapper.toMovieVO(movie, ViewFeature.SEARCHED) }
                         .map { movieVO ->
                             MovieItem(movieVO) {
                                 openMovieDetails(
